* ep-engine/watson:
  [BP] MB-29531: Replace revSeqno with a 48-bit counter

+++ src/ep.cc
@@ -3201,7 +3201,7 @@ int EventuallyPersistentStore::flushVBucket(uint16_t vbid) {
                     vbstate.maxCas = std::max(vbstate.maxCas, item->getCas());
                     if (item->isDeleted()) {
                         vbstate.maxDeletedSeqno =
-                                std::max(vbstate.maxDeletedSeqno,
+                                std::max(uint64_t(vbstate.maxDeletedSeqno),
                                          item->getRevSeqno());
                     }
                     ++stats.flusher_todo;
